# contract/openapi.yaml
openapi: 3.1.0
info:
  title: SaaS Starter API
  version: 1.0.0
  description: |
    Shared API contract for the SaaS Starter Kit.
    All backend implementations (Go, Python, Node.js, etc.) must conform to this spec.
servers:
  - url: http://localhost:8080
    description: Default development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      required: [id, email, role, verified, two_fa_enabled]
      properties:
        id:
          type: string
          description: Public UUID (never the internal serial ID)
        username:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [user, admin]
        language:
          type: string
        country:
          type: string
        address:
          type: string
        phone:
          type: string
        verified:
          type: boolean
        two_fa_enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TokenPair:
      type: object
      required: [access_token, refresh_token, expires_in, token_type]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
        token_type:
          type: string
          default: Bearer
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        error_code:
          type: string
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            oneOf:
              - type: string
              - type: array
                items:
                  type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── Auth ──────────────────────────────────────────────
...
    LoginResponse:
      type: object
      required: [message, user, access_token, refresh_token, expires_in, token_type]
      properties:
        message:
          type: string
        user:
          $ref: "#/components/schemas/User"
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
        token_type:
          type: string
          default: Bearer
...
    Verify2FARequest:
      type: object
      required: [user_id, code]
      properties:
        user_id:
          type: string
        code:
          type: string
    Resend2FARequest:
      type: object
      required: [user_id]
      properties:
        user_id:
          type: string
    VerifyTOTPLoginRequest:
      type: object
      required: [user_id, code]
      properties:
        user_id:
          type: string
        code:
          type: string
        remember_me:
          type: boolean
          default: false
...
    Setup2FAResponse:
      type: object
      required: [otpauth_url, secret, message]
      properties:
        otpauth_url:
          type: string
        secret:
          type: string
        message:
          type: string
    Verify2FASetupRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
    Verify2FASetupResponse:
      type: object
      required: [message, recovery_codes, warning]
      properties:
        message:
          type: string
        recovery_codes:
          type: array
          items:
            type: string
        warning:
          type: string
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
...
paths:
...
  /auth/refresh-token:
    post:
      tags: [Auth]
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/verify-2fa:
    post:
      tags: [Auth]
      summary: Verify email-based 2FA code
      operationId: verify2FA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Verify2FARequest"
      responses:
        "200":
          description: 2FA verified, tokens returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/resend-2fa:
    post:
      tags: [Auth]
      summary: Resend email-based 2FA code
      operationId: resend2FA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Resend2FARequest"
      responses:
        "200":
          description: Code resent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/verify-totp-login:
    post:
      tags: [Auth]
      summary: Verify TOTP code during login
      operationId: verifyTOTPLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyTOTPLoginRequest"
      responses:
        "200":
          description: TOTP verified, tokens returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid TOTP code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/2fa/setup:
    post:
      tags: [Auth (Protected)]
      summary: Start TOTP 2FA setup
      operationId: setup2FA
      security:
        - bearerAuth: []
      responses:
        "200":
          description: QR code and secret returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Setup2FAResponse"
  /auth/2fa/verify-setup:
    post:
      tags: [Auth (Protected)]
      summary: Confirm TOTP setup with a valid code
      operationId: verify2FASetup
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Verify2FASetupRequest"
      responses:
        "200":
          description: 2FA enabled, recovery codes returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Verify2FASetupResponse"
        "400":
          description: Invalid TOTP code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── User Profile ──────────────────────────────────────
...
  /api/admin/blocked-ips:
    get:
      tags: [Admin - Blocked IPs]
      summary: List all blocked IPs
      operationId: getBlockedIPs
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Blocked IP list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    ip:
                      type: string
                    blocked_until:
                      type: string
                      format: date-time
  /api/admin/blocked-ips/{ip}:
    delete:
      tags: [Admin - Blocked IPs]
      summary: Unblock a specific IP
      operationId: unblockIP
      security:
        - bearerAuth: []
      parameters:
        - name: ip
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: IP unblocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Admin - Settings ──────────────────────────────────
...